requirements.txt
streamlit==1.44.1
ibm-watsonx-ai==1.3.3
ibm-cloud-sdk-core==3.23.0
python-dotenv==1.1.0
chromadb==1.0.0
sentence-transformers==4.0.2
transformers==4.50.3